Adding more context to async shard
