+++ src/Marten/Events/Daemon/New/GroupExecution.cs
@@ -1,0 +1,18 @@
+using System.Threading;
+using System.Threading.Tasks;
+using Marten.Events.Projections;
+using Marten.Storage;
+
+namespace Marten.Events.Daemon.New;
+
+internal record GroupExecution(
+    IProjectionSource Source,
+    EventRange Range,
+    IMartenDatabase Database,
+    DocumentStore Store)
+{
+    public ValueTask<EventRangeGroup> GroupAsync(CancellationToken token)
+    {
+        return Source.GroupEvents(Store, Database, Range, token);
+    }
+}
