+++ src/Marten/Events/Daemon/New/ISubscriptionAgent.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Marten.Events.Daemon.New;
+
+public interface ISubscriptionAgent
+{
+    void Pause(TimeSpan time);
+    void MarkSuccess(long processedCeiling);
+}
