+++ src/Marten/Events/Daemon/New/ISubscriptionExecution.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Marten.Events.Daemon.New;
+
+public interface ISubscriptionExecution: IAsyncDisposable
+{
+    void Enqueue(EventPage page, ISubscriptionAgent subscriptionAgent);
+}
